Fixing issues with files not being removed during revert operation.
git-svn-id: https://svn.code.sf.net/p/tsc-tpm/code@255 af914a7b-af0d-0410-90d3-cea0664cafdb

+++ distribution/installation/remove-action.rb
@@ -69,8 +69,8 @@ module Installation
     end
 
     def make_target(progress, logger)
-      return unless File.exist?(target)
-      return unless if_types.empty? == false && if_types.include?(File.ftype(target))
+      return unless File.symlink?(target) || File.exist?(target)
+      return unless if_types.empty? || if_types.include?(File.ftype(target))
 
       FileUtils.remove_entry target
     end
